Time: 4 ms (33.69%), Space: 5.4 MB (60.20%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.c
@@ -2,7 +2,7 @@ int scoreOfString(char* s) {
     int n = strlen(s);
     int sum = 0;
     for (int i = 0; i < n - 1; i++){
-        sum += abs((int)s[i] - (int)s[i+1]);
+        sum += abs(s[i] - s[i+1]);
     }
     return sum;
 }
